Mixnode and gateway bonding tests moved
